Reworking game-object. Commented other things.
I think the general idea of what the game-object is is now finalized. I'm
pretty happy with this version. A main name, an aliases slot for extra
names, and an adjectives slot for adjectives. I also went through and
commented and reorganized other code for the other classes. I think I'm
mostly done here.