Renamed <door> to <exit>, fixed get-<object-type>
get-players was broken, so I fixed it. Also, added a get-exits function that
does the same, but for <exit> objects. <exit> objects weren't there before,
since I just renamed <door> to <exit>, to be more clear (I was constantly
making the mistake anyway, so it's the more natural name for it, I think).