More changes to the way commands are structured
Stopping this for now. I want to be able to test this as I develop it. I'm
gonna go ahead and figure out the rest of the system (account creation and
login, character creation and selection). Once that's set, I should be able
to just jump into a room and start figuring out various ways to do commands.