Metric assfuck of changes
Did a bunch of stuff. Added engine.lisp, which is the main file. Worked my
way up and wrote a sort-of-main function, and a main client handling thing.
Also, the server will now process user input through the parser and command
handler, although all the conditions need to be handled (it crashes when it
doesn't know a verb...)
Overall, too much work for a single commit. It's also incredibly unstable
and buggy, and makes me quite unhappy. This is gonna take a while...