Fixed bug when "say" is not a verb
Because " defaulting to "say" is a bit of a hack, it was leaking and caused
some funny errors (it snuck itself out of the parser...). This is still a
monkey patch, but I haven't figured out a cleaner way to do this. Perhaps
let the AST go out even if there's no verb, as long as there's a chat
string, and take care of funky binding elsewhere. That, or don't allow the
defaulting.