Lots of fixes to broken packages
as expected, the great rift has caused a bunch of errors all over the place.
To make matters worse, I mixed in the removal of the account and avatar
slots from <client>, which is causing a bunch of other errors, mixed with
the package errors, in the login module (and everything that uses it). Not
everything is fixed yet, but here's a bunch of stuff.