Initializing CLIENT's SOUL slot to NIL.
Client teardown checks whether the SOUL slot is bound, and cleans it up if it is. If the client
exits early (before being associated with a soul), this would error.

An alternative would be to have the cleanup also check if the slot is bound, but I prefer this
solution because it addresses the issue at a lower level.